requirements.txt
streamlit
anthropic
pillow
requests
elevenlabs==0.2.24